chapter17/requirements.txt
# requirements.txt — Chapter 17: Epilogue: The Future of Intelligent Agents
# Book: AI Agents by Imran Ahmad (Packt, 2025)
# Tested with Python 3.10, 3.11, 3.12

numpy>=1.26.0,<2.1.0
scipy>=1.13.0
python-dotenv>=1.0.0
jupyter>=1.0.0
ipykernel>=6.29.0
